requirements.txt
aeidon
asgiref
attrs
Automat
awscli
bcrypt
beautifulsoup4
blinker
botocore
Brotli
cclib
certifi
distlib
distro
Django
djangorestframework
docutils
fasteners
filelock
future
html5lib
httplib2
hyperlink
idna
incremental
iniparse
jeepney
jmespath
keyring
launchpadlib
lockfile
lxml
macaroonbakery
Mako
MarkupSafe
monotonic
mutagen
netifaces
numpy
oauthlib
olefile
paramiko
pexpect
phply
Pillow
platformdirs
ply
protobuf
ptyprocess
pyasn1
pycountry
pycryptodomex
pyparsing
pytz
requests
sqlparse
virtualenv
